Get rid of worker-library
